np.int to int

+++ pytools/load_grid.py
@@ -35,10 +35,10 @@ def balance_mpi_2D(n, comm_size=None):
             raise ValueError("Ny is not power of 2 (i.e. 2^m)")
 
         # print('Generating hilbert with 2^{} {}'.format(m0,m1))
-        hgen = pyrunko.tools.twoD.HilbertGen(np.int(m0), np.int(m1))
+        hgen = pyrunko.tools.twoD.HilbertGen(int(m0), int(m1))
 
-        igrid = np.zeros((nx, ny), np.int64)
-        grid = np.zeros((nx, ny))  # , np.int64)
+        igrid = np.zeros((nx, ny), int)
+        grid = np.zeros((nx, ny))  # , int)
 
         for i in range(nx):
             for j in range(ny):
@@ -93,10 +93,10 @@ def balance_mpi_3D(n, comm_size=None):
             raise ValueError("Nz is not power of 2 (i.e. 2^m)")
 
         # print('Generating hilbert with 2^{} {}'.format(m0,m1))
-        hgen = pyrunko.tools.threeD.HilbertGen(np.int(m0), np.int(m1), np.int(m2))
+        hgen = pyrunko.tools.threeD.HilbertGen(int(m0), int(m1), int(m2))
 
-        igrid = np.zeros((nx, ny, nz), np.int64)
-        grid = np.zeros((nx, ny, nz))  # , np.int64)
+        igrid = np.zeros((nx, ny, nz), int)
+        grid = np.zeros((nx, ny, nz))  # , int)
 
         for i in range(nx):
             for j in range(ny):
@@ -162,10 +162,10 @@ def load_catepillar_track_mpi(
 
         # print('Generating hilbert with 2^{} {}'.format(m0,m1))
         if conf.twoD:
-            hgen = pyrunko.tools.twoD.HilbertGen(np.int(m0), np.int(m1))
+            hgen = pyrunko.tools.twoD.HilbertGen(int(m0), int(m1))
         elif conf.threeD:
-            hgen = pyrunko.tools.threeD.HilbertGen(np.int(m0), np.int(m1), np.int(m2) )
-        grid = np.zeros((nx, ny, nz))  # , np.int64)
+            hgen = pyrunko.tools.threeD.HilbertGen(int(m0), int(m1), int(m2) )
+        grid = np.zeros((nx, ny, nz))  # , int)
 
         for i in range(nx):
             for j in range(ny):
@@ -180,7 +180,7 @@ def load_catepillar_track_mpi(
 
         # create alternating true grid
         nxt = n.get_Nx()
-        igrid = np.zeros((nxt, ny, nz), np.int64)
+        igrid = np.zeros((nxt, ny, nz), int)
 
         for i in range(nxt):
             for j in range(ny):
@@ -227,10 +227,10 @@ def load_mpi_randomly(n):
 #load nodes to be in stripe formation (splitted in X=horizontal direction)
 def load_mpi_x_strides(n, conf):
     if n.rank() == 0: #only master initializes; then sends
-        stride = np.zeros( (n.get_Nx()), np.int64)
+        stride = np.zeros( (n.get_Nx()), int)
         dx = np.float(n.get_Nx()) / np.float(n.size() ) 
         for i in range(n.get_Nx()):
-            val = np.int( i/dx )
+            val = int( i/dx )
             stride[i] = val
 
         for j in range(n.get_Ny()):
@@ -244,10 +244,10 @@ def load_mpi_x_strides(n, conf):
 #load nodes to be in stripe formation (splitted in Y=vertical direction)
 def load_mpi_y_strides(n, conf):
     if n.rank() == 0: #only master initializes; then sends
-        stride = np.zeros( (n.get_Ny()), np.int64)
+        stride = np.zeros( (n.get_Ny()), int)
         dy = np.float(n.get_Ny()) / np.float(n.size() ) 
         for j in range(n.get_Ny()):
-            val = np.int( j/dy )
+            val = int( j/dy )
             stride[j] = val
 
         for i in range(n.get_Nx()):
@@ -267,9 +267,9 @@ def get_mpi_grid(grid, conf):
     ny = grid.get_Ny()
     nz = grid.get_Nz()
 
-    #mpi_grid = np.zeros((nx, ny, nz), np.int64)
-    #mpi_grid = np.zeros((ny, nx, nz), np.int64)
-    mpi_grid = np.zeros((nx, ny, nz), np.int64)
+    #mpi_grid = np.zeros((nx, ny, nz), int64)
+    #mpi_grid = np.zeros((ny, nx, nz), int64)
+    mpi_grid = np.zeros((nx, ny, nz), int)
 
     for i in range(nx):
         for j in range(ny):
@@ -294,7 +294,7 @@ def get_tile_id_grid(grid, conf):
     nx = grid.get_Nx()
     ny = grid.get_Ny()
     nz = grid.get_Nz()
-    tile_grid = np.zeros( (nx,ny,nz), np.int64 )
+    tile_grid = np.zeros( (nx,ny,nz), int )
 
     for cid in grid.get_tile_ids():
         tile = grid.get_tile( cid )
